sanitize header key and values
